* origin/develop: (30 commits)
  Remove a test focus
  Fixed test cases
  Fix podspec not updated with latest version number
  Update for v4.5.0
  Follow style guide for boolean check
  Updated formatting of README.md
  Added steps for using the lock screen manager from SDL's bundle to the readme.
  Added check when trying to upload an appIcon for a head unit that does not support graphics.
  Update for v4.5.0
  Fixed issue with trying to upload a file that is nil will cause an infinite loop.
  Fixed an issue with initializer causing infinite loop.
  Improved readability for nil-checks.
  Fixed issue relating to SDLLocationDetails' searchAddress using the wrong key for object store.
  Fixed formatting for some interfaces.
  Fixed issue with LocationDetails' properties not properly becoming initialized.
  Fixed issue with waypoints not being correctly initialized as SDLLocationDetails
  Adding in nil checks for completion handlers of Streaming Media Manager.
  Updated description to return correct control data name.
  Update changelog for v4.5.0-beta.1
  Update for v4.5.0-beta.1 release
  ...

# Conflicts:
#	SmartDeviceLink/SDLGetWaypoints.h
#	SmartDeviceLink/SDLGetWaypoints.m
#	SmartDeviceLink/SDLGetWaypointsResponse.h
#	SmartDeviceLink/SDLGetWaypointsResponse.m
#	SmartDeviceLink/SDLLocationDetails.m
#	SmartDeviceLink/SDLOnWaypointChange.h
#	SmartDeviceLink/SDLOnWaypointChange.m
#	SmartDeviceLink/SDLPerformInteraction.h
#	SmartDeviceLink/SDLPerformInteraction.m
#	SmartDeviceLink/SDLShowConstantTBT.h
#	SmartDeviceLink/SDLShowConstantTBT.m
#	SmartDeviceLink/SDLSoftButton.h
#	SmartDeviceLink/SDLSubscribeWaypoints.h
#	SmartDeviceLink/SDLSubscribeWaypoints.m
#	SmartDeviceLink/SDLSubscribeWaypointsResponse.h
#	SmartDeviceLink/SDLUnsubscribeWaypoints.h
#	SmartDeviceLink/SDLUnsubscribeWaypoints.m
#	SmartDeviceLink/SDLUnsubscribeWaypointsResponse.h
#	SmartDeviceLink/SDLUnsubscribeWaypointsResponse.m

+++ SmartDeviceLink/SDLAddSubMenu.h
@@ -21,9 +21,9 @@ NS_ASSUME_NONNULL_BEGIN
 
 @interface SDLAddSubMenu : SDLRPCRequest
 
-- (instancetype)initWithId:(UInt8)menuId menuName:(NSString *)menuName;
+- (instancetype)initWithId:(UInt32)menuId menuName:(NSString *)menuName;
 
-- (instancetype)initWithId:(UInt8)menuId menuName:(NSString *)menuName position:(UInt8)position;
+- (instancetype)initWithId:(UInt32)menuId menuName:(NSString *)menuName position:(UInt8)position;
 
 /**
  * @abstract a Menu ID that identifies a sub menu
