+++ SmartDeviceLink/SDLSetMediaClockTimer.m
@@ -18,7 +18,7 @@ NS_ASSUME_NONNULL_BEGIN
 }
 
 
-- (instancetype)initWithUpdateMode:(SDLUpdateMode)updateMode hours:(NSInteger)hours minutes:(NSInteger)minutes seconds:(NSInteger)seconds {
+- (instancetype)initWithUpdateMode:(SDLUpdateMode)updateMode hours:(UInt8)hours minutes:(UInt8)minutes seconds:(UInt8)seconds {
     self = [self initWithUpdateMode:updateMode];
     if (!self) {
         return nil;
